pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personal-agent"
version = "0.1.0"
description = "A personal AI agent for email, calendar, banking, and media management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Alois Wirth", email = "alois_wirth@gmx.de" }
]
keywords = ["agent", "ai", "langgraph", "personal-assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "caldav>=1.3.0",
    "icalendar>=5.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3.0",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
s3 = [
    "boto3>=1.34",
]
api = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
]
transcription = [
    "openai-whisper>=20231117",
    # or for API-based:
    # "openai>=1.0",
]
all = [
    "personal-agent[dev,s3,api,transcription]",
]

[project.scripts]
personal-agent = "src.cli:main"

[project.urls]
Homepage = "https://github.com/aloiswirth/personal_agent"
Repository = "https://github.com/aloiswirth/personal_agent"
Documentation = "https://github.com/aloiswirth/personal_agent/docs"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
